feat: migrate aggregator plugins to new sample config format (#10912)
