+++ plugins/aggregators/merge/README.md
@@ -10,6 +10,7 @@ be handled more efficiently by the output.
 ## Configuration
 
 ```toml
+# Merge metrics into multifield metrics by series key
 [[aggregators.merge]]
   ## If true, the original metric will be dropped by the
   ## aggregator and will not get sent to the output plugins.
