+++ plugins/aggregators/starlark/starlark.go
@@ -7,39 +7,6 @@ import (
 	"go.starlark.net/starlark"
 )
 
-const (
-	description  = "Aggregate metrics using a Starlark script"
-	sampleConfig = `
-  ## The Starlark source can be set as a string in this configuration file, or
-  ## by referencing a file containing the script.  Only one source or script
-  ## should be set at once.
-  ##
-  ## Source of the Starlark script.
-  source = '''
-state = {}
-
-def add(metric):
-  state["last"] = metric
-
-def push():
-  return state.get("last")
-
-def reset():
-  state.clear()
-'''
-
-  ## File containing a Starlark script.
-  # script = "/usr/local/bin/myscript.star"
-
-  ## The constants of the Starlark script.
-  # [aggregators.starlark.constants]
-  #   max_size = 10
-  #   threshold = 0.75
-  #   default_name = "Julia"
-  #   debug_mode = true
-`
-)
-
 type Starlark struct {
 	common.StarlarkCommon
 }
@@ -70,14 +37,6 @@ func (s *Starlark) Init() error {
 	}
 
 	return nil
-}
-
-func (s *Starlark) SampleConfig() string {
-	return sampleConfig
-}
-
-func (s *Starlark) Description() string {
-	return description
 }
 
 func (s *Starlark) Add(metric telegraf.Metric) {
