+++ plugins/aggregators/valuecounter/README.md
@@ -18,6 +18,7 @@ limited set of values.
 ## Configuration
 
 ```toml
+# Count the occurrence of values in fields.
 [[aggregators.valuecounter]]
   ## General Aggregator Arguments:
   ## The period on which to flush & clear the aggregator.
